fix GPS on PG1301
The GPS module on my PG1301 does not seem to be an u-blox.
Activating ubx7 causes current time to be zero,
which in the following triggers another bug in the beacon queuing.
(After the first beacon was sent, infinite number of beacon messages
are added to the jit_queue, and discarded)